Resolve 2 hacking problems

+++ Mathematics/Candy Lottery.cpp
@@ -2,13 +2,13 @@
 #include <iomanip>
 using namespace std;
 int n, k;
-long double res;
+double res;
  
 int main()
 {
     cin >> n >> k;
     for(int i=1; i<=k; ++i) {
-        long double x = 1, y = 1;
+        double x = 1, y = 1;
         for(int j=1; j<=n; ++j) {
             x = x * i / k;
             y = y * (i - 1) / k;
